♻️ Refactor code
