+++ assets/icons/addlink.tsx
@@ -1,30 +1,28 @@
 export default function AddLink(): JSX.Element {
   return (
-    <>
-      <svg
-        width={52}
-        height={52}
-        viewBox="0 0 52 52"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <circle
-          cx={26}
-          cy={26}
-          r={24.5}
-          stroke="#4F4F4F"
-          strokeWidth={3}
-          strokeLinecap="square"
-          strokeDasharray="12 10"
-        />
-        <path
-          d="M26.252 16.66v19.143M16.66 25.748h19.142"
-          stroke="#4F4F4F"
-          strokeWidth={4}
-          strokeLinecap="round"
-          strokeLinejoin="round"
-        />
-      </svg>
-    </>
+    <svg
+      width={52}
+      height={52}
+      viewBox="0 0 52 52"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <circle
+        cx={26}
+        cy={26}
+        r={24.5}
+        stroke="#4F4F4F"
+        strokeWidth={3}
+        strokeLinecap="square"
+        strokeDasharray="12 10"
+      />
+      <path
+        d="M26.252 16.66v19.143M16.66 25.748h19.142"
+        stroke="#4F4F4F"
+        strokeWidth={4}
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </svg>
   );
 }
