+++ assets/icons/circle.tsx
@@ -1,15 +1,13 @@
 export default function Circle(): JSX.Element {
   return (
-    <>
-      <svg
-        width={51}
-        height={77}
-        viewBox="0 0 51 77"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <circle cx="38.5" cy="38.5" r={35} stroke="#56ccf2" strokeWidth={7} />
-      </svg>
-    </>
+    <svg
+      width={51}
+      height={77}
+      viewBox="0 0 51 77"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <circle cx="38.5" cy="38.5" r={35} stroke="#56ccf2" strokeWidth={7} />
+    </svg>
   );
 }
