--- assets/icons/copyclipboard.tsx
@@ -1,32 +1,0 @@
-export default function CopyClipboard(): JSX.Element {
-    return (
-        <>
-            <svg
-                width={13}
-                height={18}
-                viewBox="0 0 12 16"
-                fill="none"
-                xmlns="http://www.w3.org/2000/svg"
-            >
-                <mask id="prefix__a" fill="#fff">
-                    <path
-                        fillRule="evenodd"
-                        clipRule="evenodd"
-                        d="M6 2H2v7.2h1.2v2H2a2 2 0 01-2-2V2a2 2 0 012-2h4a2 2 0 012 2v1.446H6V2zm0 11.2v-7h4v7H6zm-2-7a2 2 0 012-2h4a2 2 0 012 2v7a2 2 0 01-2 2H6a2 2 0 01-2-2v-7z"
-                    />
-                </mask>
-                <path
-                    fillRule="evenodd"
-                    clipRule="evenodd"
-                    d="M6 2H2v7.2h1.2v2H2a2 2 0 01-2-2V2a2 2 0 012-2h4a2 2 0 012 2v1.446H6V2zm0 11.2v-7h4v7H6zm-2-7a2 2 0 012-2h4a2 2 0 012 2v7a2 2 0 01-2 2H6a2 2 0 01-2-2v-7z"
-                    fill="#C3C3C3"
-                />
-                <path
-                    d="M2 2V0H0v2h2zm4 0h2V0H6v2zM2 9.2H0v2h2v-2zm1.2 0h2v-2h-2v2zm0 2v2h2v-2h-2zM8 3.446v2h2v-2H8zm-2 0H4v2h2v-2zM6 6.2v-2H4v2h2zm0 7H4v2h2v-2zm4-7h2v-2h-2v2zm0 7v2h2v-2h-2zM2 4h4V0H2v4zm2 5.2V2H0v7.2h4zm-.8-2H2v4h1.2v-4zm2 4v-2h-4v2h4zm-3.2 2h1.2v-4H2v4zm-4-4a4 4 0 004 4v-4h-4zM-2 2v7.2h4V2h-4zm4-4a4 4 0 00-4 4h4v-4zm4 0H2v4h4v-4zm4 4a4 4 0 00-4-4v4h4zm0 1.446V2H6v1.446h4zm-4 2h2v-4H6v4zM4 2v1.446h4V2H4zm0 4.2v7h4v-7H4zm6-2H6v4h4v-4zm2 9v-7H8v7h4zm-6 2h4v-4H6v4zm0-13a4 4 0 00-4 4h4v-4zm4 0H6v4h4v-4zm4 4a4 4 0 00-4-4v4h4zm0 7v-7h-4v7h4zm-4 4a4 4 0 004-4h-4v4zm-4 0h4v-4H6v4zm-4-4a4 4 0 004 4v-4H2zm0-7v7h4v-7H2z"
-                    fill="#C3C3C3"
-                    mask="url(#prefix__a)"
-                />
-            </svg>
-        </>
-    );
-}
