+++ assets/icons/doublecircle.tsx
@@ -1,16 +1,14 @@
 export default function DoubleCircle(): JSX.Element {
   return (
-    <>
-      <svg
-        width={70}
-        height={97}
-        viewBox="0 0 70 97"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <circle cx="11.5" cy="43.5" r={35} stroke="#56ccf2" strokeWidth={7} />
-        <circle cx="21.5" cy="48.5" r={45} stroke="#56ccf2" strokeWidth={7} />
-      </svg>
-    </>
+    <svg
+      width={70}
+      height={97}
+      viewBox="0 0 70 97"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <circle cx="11.5" cy="43.5" r={35} stroke="#56ccf2" strokeWidth={7} />
+      <circle cx="21.5" cy="48.5" r={45} stroke="#56ccf2" strokeWidth={7} />
+    </svg>
   );
 }
