+++ assets/icons/editpencil.tsx
@@ -1,19 +1,17 @@
 export default function EditPencil(): JSX.Element {
-    return (
-        <>
-            <svg
-                width={13}
-                height={13}
-                viewBox="0 0 13 13"
-                fill="none"
-                xmlns="http://www.w3.org/2000/svg"
-            >
-                <path
-                    d="M1 9.857L8.857 2 11 4.143 3.143 12H1V9.857z"
-                    stroke="#C4C3C3"
-                    strokeWidth={2}
-                />
-            </svg>
-        </>
-    );
+  return (
+    <svg
+      width={13}
+      height={13}
+      viewBox="0 0 13 13"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M1 9.857L8.857 2 11 4.143 3.143 12H1V9.857z"
+        stroke="#C4C3C3"
+        strokeWidth={2}
+      />
+    </svg>
+  );
 }
