+++ assets/icons/ellipsegray.tsx
@@ -1,15 +1,13 @@
 export default function EllipseGray(): JSX.Element {
-    return (
-        <>
-            <svg
-                width={358}
-                height={307}
-                viewBox="0 0 358 307"
-                fill="none"
-                xmlns="http://www.w3.org/2000/svg"
-            >
-                <circle cx={316} cy={-9} r={316} fill="#F2F2F2" fillOpacity={0.4} />
-            </svg>
-        </>
-    );
+  return (
+    <svg
+      width={358}
+      height={307}
+      viewBox="0 0 358 307"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <circle cx={316} cy={-9} r={316} fill="#F2F2F2" fillOpacity={0.4} />
+    </svg>
+  );
 }
