+++ assets/icons/index.tsx
@@ -2,7 +2,6 @@ export { default as Loading } from "./loading";
 export { default as NewWindow } from "./newWindow";
 export { default as NoLinks } from "./nolinks";
 export { default as Pencil } from "./pencil";
-export { default as Spike } from "./spike";
 export { default as Tick } from "./tick";
 export { default as Trash } from "./trash";
 export { default as Logo } from "./logo";
@@ -17,7 +16,6 @@ export { default as Circle } from "./circle";
 export { default as DoubleCircle } from "./doublecircle";
 export { default as EditPencil } from "./editpencil";
 export { default as LinksLogoBg } from "./linkslogobg";
-export { default as CopyClipboard } from "./copyclipboard";
 export { default as EllipseGray } from "./ellipsegray";
 export { default as AddLink } from "./addlink";
 export { default as Info } from "./info";
