+++ assets/icons/info.tsx
@@ -1,18 +1,16 @@
 export default function Info(): JSX.Element {
   return (
-    <>
-      <svg
-        width={15}
-        height={15}
-        viewBox="0 0 11 12"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <path
-          d="M6.05 3.92h-1.1V2.805h1.1v1.113zm0 4.454h-1.1V5.033h1.1v3.34zm-.545 2.784C8.541 11.158 11 8.663 11 5.59 11 2.516 8.541.022 5.505.022 2.464.022 0 2.516 0 5.59c0 3.073 2.464 5.568 5.505 5.568zM5.5 1.135c2.431 0 4.4 1.994 4.4 4.455 0 2.461-1.969 4.454-4.4 4.454S1.1 8.051 1.1 5.59c0-2.461 1.969-4.455 4.4-4.455z"
-          fill="#A7A6A6"
-        />
-      </svg>
-    </>
+    <svg
+      width={15}
+      height={15}
+      viewBox="0 0 11 12"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M6.05 3.92h-1.1V2.805h1.1v1.113zm0 4.454h-1.1V5.033h1.1v3.34zm-.545 2.784C8.541 11.158 11 8.663 11 5.59 11 2.516 8.541.022 5.505.022 2.464.022 0 2.516 0 5.59c0 3.073 2.464 5.568 5.505 5.568zM5.5 1.135c2.431 0 4.4 1.994 4.4 4.455 0 2.461-1.969 4.454-4.4 4.454S1.1 8.051 1.1 5.59c0-2.461 1.969-4.455 4.4-4.455z"
+        fill="#A7A6A6"
+      />
+    </svg>
   );
 }
