+++ assets/icons/linkslogobg.tsx
@@ -1,34 +1,32 @@
 export default function LinksLogoBg(): JSX.Element {
   return (
-    <>
-      <svg
-        width={428}
-        height={272}
-        viewBox="0 0 428 272"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <path
-          d="M313.729 396.484c-14.687 18.207-34.422 26.072-57.151 26.706-38.577 1.003-65.368-22.305-71.933-62.083-2.08-12.625-3.476-25.356-5.253-38.532l38.218 21.939c.96 7.125 1.446 14.087 4.406 20.637 11.739 25.961 46.616 31.296 64.867 9.601 7.075-8.416 9.081-18.764 7.985-29.407a26423.784 26423.784 0 00-17.679-164.066c-1.336-11.978-.393-24.017-2.083-35.973-1.818-12.796-5.213-24.883-12.789-35.553a9.59 9.59 0 01-.653-1.401 33.484 33.484 0 0110.14 3.092l7.36 4.225a50.82 50.82 0 0118.222 18.327c8.299 13.999 12.082 29.463 13.851 45.384a27737.187 27737.187 0 0117.454 161.657c1.496 14.082-.869 27.74-6.535 40.768-.84 1.632-2.321 2.971-2.203 5.029l-2.696 4.698c-1.811 1.211-2.461 3.225-3.528 4.952z"
-          fill="#F2F2F2"
-          fillOpacity={0.8}
-        />
-        <path
-          d="M279.389 115.217c-.085.15-.162.304-.231.462l-7.36-4.225c.082-.144.165-.288.276-.423l7.315 4.186zM-8.198 222.383c-10.424 9.552-17.127 21.191-21.31 35.153a64.227 64.227 0 01-6.013-15.417c-3.809-15.717-4.735-31.353 2.269-46.486 8.622-18.596 22.79-31.772 41.277-40.122a25634.593 25634.593 0 01146.508-65.583c14.368-6.384 29.347-10.913 45.3-10.77 25.372.22 50.39 19.213 59.1 44.358 5.63 16.264 8.067 32.678 3.146 49.534-5.647 19.374-19.065 32.46-35.523 42.939-3.455 2.203-7.156 4.025-10.753 6.01-.572.223-1.159.405-1.756.545l-4.554-35.449c9.341-6.555 15.65-15.185 18.178-26.122 3.686-15.957-2.6-28.958-14.05-39.279-10.656-9.607-23.624-8.774-36.06-4.003-28.573 10.912-55.947 24.49-83.923 36.805-25.138 11.065-50.12 22.396-75.158 33.682-6.408 2.889-12.474 6.398-17.769 11.116-6.842 6.1-10.9 13.279-8.91 23.089z"
-          fill="#F2F2F2"
-          fillOpacity={0.8}
-        />
-        <path
-          d="M231.844 432.968c-12.583 14.23-30.852 16.787-49.508 6.632-17.166-9.342-34.19-18.943-51.207-28.555-23.412-13.208-46.924-26.253-70.095-39.87-18.306-10.775-37.18-20.707-54.405-33.188-39.868-28.9-36.726-70.411-16.113-99.677 27.79-39.457 67.727-36.129 98.236-20.133 3.193 1.675 6.254 3.59 10.192 5.85a71172.717 71172.717 0 01-37.808 16.125c-12.374-4.552-26.665.077-37.419 12.039-18.253 20.305-13.69 48.44 10.039 62.062 47.076 27.025 94.146 54.057 141.209 81.097 5.471 3.141 10.771 6.594 16.124 9.94a5.37 5.37 0 011.743 1.555c8.924 14.256 22.198 22.128 38.27 25.746.264.089.514.216.742.377z"
-          fill="#F2F2F2"
-          fillOpacity={0.8}
-        />
-        <path
-          d="M122.832 277.081l38.635-17.34c26.612 16.324 54.827 30.059 80.64 47.88a92.1 92.1 0 0120.121 18.775c11.333 14.4 14.088 28.281 9.077 45.974-9.456 5.982-23.903 3.825-34.507-5.204 4.993-16.459-5.35-25.646-16.938-33.475-22.287-15.072-46.38-27.152-69.522-40.819-8.88-5.249-17.896-10.274-27.506-15.791zM156.75 209.962c1.232 12.446 2.44 24.659 3.636 36.88-16.081 8.499-32.545 15.363-48.843 22.581-18.777 8.307-37.522 16.7-56.663 24.175-9.836 3.842-18.398 4.392-27.603 1.591-5.828-9.064-4.583-23.727 3.002-32.911 8.396 4.341 14.642-2.211 21.323-5.135 35.041-15.36 69.864-31.256 105.148-47.181zM210.071 273.6l-38.017-21.824c-.998-10.184-2.084-20.535-3.031-30.889-1.654-17.934-4.262-35.801-2.414-53.922 2.045-20.073 11.508-34.793 29.865-43.477 14.619 4.684 19.893 17.179 19.708 24.536-14.907 4.748-16.29 11.855-15.787 26.01 1.055 28.967 5.623 57.618 8.303 86.437.39 4.185.868 8.358 1.373 13.129z"
-          fill="#F2F2F2"
-          fillOpacity={0.8}
-        />
-      </svg>
-    </>
+    <svg
+      width={428}
+      height={272}
+      viewBox="0 0 428 272"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M313.729 396.484c-14.687 18.207-34.422 26.072-57.151 26.706-38.577 1.003-65.368-22.305-71.933-62.083-2.08-12.625-3.476-25.356-5.253-38.532l38.218 21.939c.96 7.125 1.446 14.087 4.406 20.637 11.739 25.961 46.616 31.296 64.867 9.601 7.075-8.416 9.081-18.764 7.985-29.407a26423.784 26423.784 0 00-17.679-164.066c-1.336-11.978-.393-24.017-2.083-35.973-1.818-12.796-5.213-24.883-12.789-35.553a9.59 9.59 0 01-.653-1.401 33.484 33.484 0 0110.14 3.092l7.36 4.225a50.82 50.82 0 0118.222 18.327c8.299 13.999 12.082 29.463 13.851 45.384a27737.187 27737.187 0 0117.454 161.657c1.496 14.082-.869 27.74-6.535 40.768-.84 1.632-2.321 2.971-2.203 5.029l-2.696 4.698c-1.811 1.211-2.461 3.225-3.528 4.952z"
+        fill="#F2F2F2"
+        fillOpacity={0.8}
+      />
+      <path
+        d="M279.389 115.217c-.085.15-.162.304-.231.462l-7.36-4.225c.082-.144.165-.288.276-.423l7.315 4.186zM-8.198 222.383c-10.424 9.552-17.127 21.191-21.31 35.153a64.227 64.227 0 01-6.013-15.417c-3.809-15.717-4.735-31.353 2.269-46.486 8.622-18.596 22.79-31.772 41.277-40.122a25634.593 25634.593 0 01146.508-65.583c14.368-6.384 29.347-10.913 45.3-10.77 25.372.22 50.39 19.213 59.1 44.358 5.63 16.264 8.067 32.678 3.146 49.534-5.647 19.374-19.065 32.46-35.523 42.939-3.455 2.203-7.156 4.025-10.753 6.01-.572.223-1.159.405-1.756.545l-4.554-35.449c9.341-6.555 15.65-15.185 18.178-26.122 3.686-15.957-2.6-28.958-14.05-39.279-10.656-9.607-23.624-8.774-36.06-4.003-28.573 10.912-55.947 24.49-83.923 36.805-25.138 11.065-50.12 22.396-75.158 33.682-6.408 2.889-12.474 6.398-17.769 11.116-6.842 6.1-10.9 13.279-8.91 23.089z"
+        fill="#F2F2F2"
+        fillOpacity={0.8}
+      />
+      <path
+        d="M231.844 432.968c-12.583 14.23-30.852 16.787-49.508 6.632-17.166-9.342-34.19-18.943-51.207-28.555-23.412-13.208-46.924-26.253-70.095-39.87-18.306-10.775-37.18-20.707-54.405-33.188-39.868-28.9-36.726-70.411-16.113-99.677 27.79-39.457 67.727-36.129 98.236-20.133 3.193 1.675 6.254 3.59 10.192 5.85a71172.717 71172.717 0 01-37.808 16.125c-12.374-4.552-26.665.077-37.419 12.039-18.253 20.305-13.69 48.44 10.039 62.062 47.076 27.025 94.146 54.057 141.209 81.097 5.471 3.141 10.771 6.594 16.124 9.94a5.37 5.37 0 011.743 1.555c8.924 14.256 22.198 22.128 38.27 25.746.264.089.514.216.742.377z"
+        fill="#F2F2F2"
+        fillOpacity={0.8}
+      />
+      <path
+        d="M122.832 277.081l38.635-17.34c26.612 16.324 54.827 30.059 80.64 47.88a92.1 92.1 0 0120.121 18.775c11.333 14.4 14.088 28.281 9.077 45.974-9.456 5.982-23.903 3.825-34.507-5.204 4.993-16.459-5.35-25.646-16.938-33.475-22.287-15.072-46.38-27.152-69.522-40.819-8.88-5.249-17.896-10.274-27.506-15.791zM156.75 209.962c1.232 12.446 2.44 24.659 3.636 36.88-16.081 8.499-32.545 15.363-48.843 22.581-18.777 8.307-37.522 16.7-56.663 24.175-9.836 3.842-18.398 4.392-27.603 1.591-5.828-9.064-4.583-23.727 3.002-32.911 8.396 4.341 14.642-2.211 21.323-5.135 35.041-15.36 69.864-31.256 105.148-47.181zM210.071 273.6l-38.017-21.824c-.998-10.184-2.084-20.535-3.031-30.889-1.654-17.934-4.262-35.801-2.414-53.922 2.045-20.073 11.508-34.793 29.865-43.477 14.619 4.684 19.893 17.179 19.708 24.536-14.907 4.748-16.29 11.855-15.787 26.01 1.055 28.967 5.623 57.618 8.303 86.437.39 4.185.868 8.358 1.373 13.129z"
+        fill="#F2F2F2"
+        fillOpacity={0.8}
+      />
+    </svg>
   );
 }
