+++ assets/icons/loading.tsx
@@ -1,8 +1,8 @@
 export default function Loading(): JSX.Element {
   return (
-    <>
+    <div className="animate-spin flex items-center justify-center">
       <svg
-        className="animate-spin w-5 text-red-400"
+        className="w-5 text-red-400"
         xmlns="http://www.w3.org/2000/svg"
         fill="none"
         viewBox="0 0 24 24"
@@ -21,6 +21,6 @@ export default function Loading(): JSX.Element {
           d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
         />
       </svg>
-    </>
+    </div>
   );
 }
