+++ assets/icons/notfoundleft.tsx
@@ -1,31 +1,29 @@
 export default function NotFoundLeft(): JSX.Element {
-    return (
-        <>
-            <svg
-                width={211}
-                height={224}
-                viewBox="0 0 211 224"
-                fill="none"
-                xmlns="http://www.w3.org/2000/svg"
-            >
-                <path
-                    d="M120.89 214.237l-7.922 1.113 1.113 7.922 7.923-1.113-1.114-7.922zm-5.698-40.542l7.923-1.113-1.114-7.922-7.922 1.113 1.113 7.922zm-100.825 14.17l-7.922 1.114 1.114 7.922 7.922-1.113-1.114-7.923zm-5.35-38.073l-7.114-3.661-1.162 2.259.353 2.515 7.922-1.113zM76.222 19.195l-1.113-7.922-4.104.577-1.896 3.685 7.113 3.66zm71.212-10.008l7.922-1.113-1.113-7.922-7.922 1.113 1.113 7.922zm16.35 116.337l-7.922 1.113 1.113 7.922 7.923-1.113-1.114-7.922zm32.433-4.558l7.923-1.114-1.114-7.922-7.922 1.113 1.113 7.923zm5.698 40.541l1.114 7.922 7.922-1.113-1.114-7.922-7.922 1.113zm-32.433 4.558l-1.113-7.922-7.923 1.114 1.114 7.922 7.922-1.114zm5.698 40.542l1.113 7.922 7.922-1.113-1.113-7.922-7.922 1.113zM55.909 140.685l-7.145-3.599-6.873 13.647 15.132-2.126-1.114-7.922zm54.291-7.63l1.113 7.922 7.783-1.094-.956-7.801-7.94.973zm-9.738-79.519l7.941-.972-.99-8.084-8.064 1.134 1.113 7.922zm-1.057.149l-1.114-7.922-4.24.595-1.856 3.859 7.21 3.468zm-5.818 12.321l7.159 3.57.05-.1.047-.1-7.256-3.37zM87.67 77.623l-7.094-3.699-.026.05-.025.05 7.145 3.599zm41.142 135.501l-5.697-40.542-15.845 2.227 5.698 40.541 15.844-2.226zm-14.733-47.351l-100.825 14.17 2.227 15.845 100.825-14.171-2.227-15.844zm-91.79 20.979l-5.35-38.074-15.845 2.227 5.351 38.074 15.845-2.227zm-6.16-33.3L83.337 22.856 69.11 15.535 1.903 146.13l14.227 7.321zM77.337 27.117L148.55 17.11l-2.227-15.844L75.11 11.273l2.226 15.844zM139.513 10.3l16.35 116.337 15.844-2.227-16.35-116.336-15.844 2.226zm25.386 123.146l32.433-4.558-2.227-15.845-32.433 4.559 2.227 15.844zm23.397-11.367l5.698 40.542 15.844-2.227-5.697-40.542-15.845 2.227zm12.507 31.506l-32.433 4.558 2.226 15.844 32.434-4.558-2.227-15.844zm-39.242 13.594l5.698 40.541 15.844-2.226-5.698-40.542-15.844 2.227zm12.506 31.506l-54.29 7.63 2.227 15.844 54.29-7.63-2.227-15.844zM57.023 148.607l54.29-7.63-2.227-15.844-54.29 7.63 2.227 15.844zm61.117-16.525l-9.737-79.518-15.882 1.944 9.738 79.519 15.881-1.945zM99.349 45.614l-1.058.149 2.227 15.844 1.058-.149-2.227-15.844zm-7.154 4.603a1004.885 1004.885 0 00-5.864 12.42l14.512 6.738c2.01-4.33 3.934-8.403 5.771-12.222l-14.419-6.936zm-5.767 12.219a719.933 719.933 0 01-5.852 11.488l14.188 7.397a738.526 738.526 0 005.982-11.744l-14.318-7.141zm-5.903 11.588l-31.76 63.062 14.29 7.197 31.76-63.062-14.29-7.197z"
-                    fill="url(#prefix__paint0_linear)"
-                />
-                <defs>
-                    <linearGradient
-                        id="prefix__paint0_linear"
-                        x1={65.988}
-                        y1={-135.526}
-                        x2={130.425}
-                        y2={322.968}
-                        gradientUnits="userSpaceOnUse"
-                    >
-                        <stop stopColor="#56CCF2" />
-                        <stop offset={1} stopColor="#6FCF97" />
-                    </linearGradient>
-                </defs>
-            </svg>
-        </>
-    );
+  return (
+    <svg
+      width={211}
+      height={224}
+      viewBox="0 0 211 224"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M120.89 214.237l-7.922 1.113 1.113 7.922 7.923-1.113-1.114-7.922zm-5.698-40.542l7.923-1.113-1.114-7.922-7.922 1.113 1.113 7.922zm-100.825 14.17l-7.922 1.114 1.114 7.922 7.922-1.113-1.114-7.923zm-5.35-38.073l-7.114-3.661-1.162 2.259.353 2.515 7.922-1.113zM76.222 19.195l-1.113-7.922-4.104.577-1.896 3.685 7.113 3.66zm71.212-10.008l7.922-1.113-1.113-7.922-7.922 1.113 1.113 7.922zm16.35 116.337l-7.922 1.113 1.113 7.922 7.923-1.113-1.114-7.922zm32.433-4.558l7.923-1.114-1.114-7.922-7.922 1.113 1.113 7.923zm5.698 40.541l1.114 7.922 7.922-1.113-1.114-7.922-7.922 1.113zm-32.433 4.558l-1.113-7.922-7.923 1.114 1.114 7.922 7.922-1.114zm5.698 40.542l1.113 7.922 7.922-1.113-1.113-7.922-7.922 1.113zM55.909 140.685l-7.145-3.599-6.873 13.647 15.132-2.126-1.114-7.922zm54.291-7.63l1.113 7.922 7.783-1.094-.956-7.801-7.94.973zm-9.738-79.519l7.941-.972-.99-8.084-8.064 1.134 1.113 7.922zm-1.057.149l-1.114-7.922-4.24.595-1.856 3.859 7.21 3.468zm-5.818 12.321l7.159 3.57.05-.1.047-.1-7.256-3.37zM87.67 77.623l-7.094-3.699-.026.05-.025.05 7.145 3.599zm41.142 135.501l-5.697-40.542-15.845 2.227 5.698 40.541 15.844-2.226zm-14.733-47.351l-100.825 14.17 2.227 15.845 100.825-14.171-2.227-15.844zm-91.79 20.979l-5.35-38.074-15.845 2.227 5.351 38.074 15.845-2.227zm-6.16-33.3L83.337 22.856 69.11 15.535 1.903 146.13l14.227 7.321zM77.337 27.117L148.55 17.11l-2.227-15.844L75.11 11.273l2.226 15.844zM139.513 10.3l16.35 116.337 15.844-2.227-16.35-116.336-15.844 2.226zm25.386 123.146l32.433-4.558-2.227-15.845-32.433 4.559 2.227 15.844zm23.397-11.367l5.698 40.542 15.844-2.227-5.697-40.542-15.845 2.227zm12.507 31.506l-32.433 4.558 2.226 15.844 32.434-4.558-2.227-15.844zm-39.242 13.594l5.698 40.541 15.844-2.226-5.698-40.542-15.844 2.227zm12.506 31.506l-54.29 7.63 2.227 15.844 54.29-7.63-2.227-15.844zM57.023 148.607l54.29-7.63-2.227-15.844-54.29 7.63 2.227 15.844zm61.117-16.525l-9.737-79.518-15.882 1.944 9.738 79.519 15.881-1.945zM99.349 45.614l-1.058.149 2.227 15.844 1.058-.149-2.227-15.844zm-7.154 4.603a1004.885 1004.885 0 00-5.864 12.42l14.512 6.738c2.01-4.33 3.934-8.403 5.771-12.222l-14.419-6.936zm-5.767 12.219a719.933 719.933 0 01-5.852 11.488l14.188 7.397a738.526 738.526 0 005.982-11.744l-14.318-7.141zm-5.903 11.588l-31.76 63.062 14.29 7.197 31.76-63.062-14.29-7.197z"
+        fill="url(#prefix__paint0_linear)"
+      />
+      <defs>
+        <linearGradient
+          id="prefix__paint0_linear"
+          x1={65.988}
+          y1={-135.526}
+          x2={130.425}
+          y2={322.968}
+          gradientUnits="userSpaceOnUse"
+        >
+          <stop stopColor="#56CCF2" />
+          <stop offset={1} stopColor="#6FCF97" />
+        </linearGradient>
+      </defs>
+    </svg>
+  );
 }
