+++ assets/icons/notfoundright.tsx
@@ -1,31 +1,29 @@
 export default function NotFoundRight(): JSX.Element {
-    return (
-        <>
-            <svg
-                width={212}
-                height={226}
-                viewBox="0 0 212 226"
-                fill="none"
-                xmlns="http://www.w3.org/2000/svg"
-            >
-                <path
-                    d="M104.38 208.821l-7.922-1.114-1.114 7.923 7.922 1.113 1.114-7.922zm5.698-40.542l7.922 1.114 1.113-7.922-7.922-1.114-1.113 7.922zM9.252 154.109l-7.922-1.113-1.113 7.922 7.922 1.113 1.113-7.922zm5.351-38.074l-5.828-5.479-1.74 1.851-.354 2.515 7.922 1.113zM115.204 9.023l1.113-7.922-4.104-.577-2.838 3.02 5.829 5.479zm71.212 10.008l7.922 1.113 1.113-7.922-7.922-1.113-1.113 7.922zm-16.35 116.337l-7.922-1.114-1.114 7.922 7.922 1.114 1.114-7.922zm32.433 4.558l7.922 1.113 1.114-7.922-7.923-1.113-1.113 7.922zm-5.698 40.541l-1.113 7.923 7.922 1.113 1.113-7.922-7.922-1.114zm-32.433-4.558l1.113-7.922-7.922-1.113-1.113 7.922 7.922 1.113zm-5.698 40.542l-1.113 7.922 7.922 1.113 1.113-7.922-7.922-1.113zm-96.48-96.244l-5.876-5.429-10.37 11.224 15.133 2.127 1.113-7.922zm54.29 7.63l-1.113 7.922 7.783 1.094 1.231-7.762-7.901-1.254zm12.558-79.123l7.901 1.255 1.277-8.043-8.064-1.134-1.114 7.922zm-1.057-.148l1.113-7.922-4.24-.596-2.847 3.197 5.974 5.32zm-8.989 10.24l5.897 5.406.076-.082.073-.085-6.046-5.239zm-8.89 9.536l-5.799-5.51-.039.04-.038.042 5.876 5.428zm2.2 141.592L118 169.393l-15.845-2.227-5.697 40.541 15.844 2.227zm-1.111-49.577l-100.825-14.17-2.227 15.844 100.825 14.17 2.227-15.844zm-94.016-5.134l5.35-38.074-15.844-2.227-5.35 38.074 15.844 2.227zm3.257-33.708l100.6-107.013-11.657-10.959-100.6 107.013 11.657 10.959zm93.658-104.57l71.212 10.008 2.227-15.844L116.317 1.1l-2.227 15.844zm64.404.973l-16.35 116.336 15.844 2.227 16.35-116.337-15.844-2.226zm-9.542 125.372l32.434 4.558 2.226-15.844-32.433-4.558-2.227 15.844zm25.625-4.477l-5.698 40.541 15.844 2.227 5.698-40.542-15.844-2.226zm3.338 33.732l-32.434-4.558-2.226 15.844 32.433 4.559 2.227-15.845zm-41.469 2.251l-5.698 40.541 15.844 2.227 5.698-40.541-15.844-2.227zm3.338 33.733l-54.291-7.63-2.227 15.844 54.291 7.63 2.227-15.844zm-98.707-80.4l54.29 7.63 2.227-15.844-54.29-7.63-2.227 15.844zm63.304.962l12.558-79.123-15.802-2.508-12.558 79.123 15.802 2.508zm5.771-88.299l-1.058-.148-2.227 15.844 1.058.149 2.227-15.845zm-8.145 2.453a991.73 991.73 0 00-9.061 10.323l12.092 10.477a988.987 988.987 0 018.917-10.158l-11.948-10.642zm-8.912 10.156a721.727 721.727 0 01-8.792 9.43l11.599 11.022a738.52 738.52 0 008.987-9.64L113.095 53.4zm-8.869 9.513l-47.912 51.864 11.752 10.857 47.913-51.864-11.753-10.857z"
-                    fill="url(#prefix__paint0_linear)"
-                />
-                <defs>
-                    <linearGradient
-                        id="prefix__paint0_linear"
-                        x1={148.012}
-                        y1={-142.526}
-                        x2={83.575}
-                        y2={315.968}
-                        gradientUnits="userSpaceOnUse"
-                    >
-                        <stop stopColor="#56CCF2" />
-                        <stop offset={1} stopColor="#6FCF97" />
-                    </linearGradient>
-                </defs>
-            </svg>
-        </>
-    );
+  return (
+    <svg
+      width={212}
+      height={226}
+      viewBox="0 0 212 226"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M104.38 208.821l-7.922-1.114-1.114 7.923 7.922 1.113 1.114-7.922zm5.698-40.542l7.922 1.114 1.113-7.922-7.922-1.114-1.113 7.922zM9.252 154.109l-7.922-1.113-1.113 7.922 7.922 1.113 1.113-7.922zm5.351-38.074l-5.828-5.479-1.74 1.851-.354 2.515 7.922 1.113zM115.204 9.023l1.113-7.922-4.104-.577-2.838 3.02 5.829 5.479zm71.212 10.008l7.922 1.113 1.113-7.922-7.922-1.113-1.113 7.922zm-16.35 116.337l-7.922-1.114-1.114 7.922 7.922 1.114 1.114-7.922zm32.433 4.558l7.922 1.113 1.114-7.922-7.923-1.113-1.113 7.922zm-5.698 40.541l-1.113 7.923 7.922 1.113 1.113-7.922-7.922-1.114zm-32.433-4.558l1.113-7.922-7.922-1.113-1.113 7.922 7.922 1.113zm-5.698 40.542l-1.113 7.922 7.922 1.113 1.113-7.922-7.922-1.113zm-96.48-96.244l-5.876-5.429-10.37 11.224 15.133 2.127 1.113-7.922zm54.29 7.63l-1.113 7.922 7.783 1.094 1.231-7.762-7.901-1.254zm12.558-79.123l7.901 1.255 1.277-8.043-8.064-1.134-1.114 7.922zm-1.057-.148l1.113-7.922-4.24-.596-2.847 3.197 5.974 5.32zm-8.989 10.24l5.897 5.406.076-.082.073-.085-6.046-5.239zm-8.89 9.536l-5.799-5.51-.039.04-.038.042 5.876 5.428zm2.2 141.592L118 169.393l-15.845-2.227-5.697 40.541 15.844 2.227zm-1.111-49.577l-100.825-14.17-2.227 15.844 100.825 14.17 2.227-15.844zm-94.016-5.134l5.35-38.074-15.844-2.227-5.35 38.074 15.844 2.227zm3.257-33.708l100.6-107.013-11.657-10.959-100.6 107.013 11.657 10.959zm93.658-104.57l71.212 10.008 2.227-15.844L116.317 1.1l-2.227 15.844zm64.404.973l-16.35 116.336 15.844 2.227 16.35-116.337-15.844-2.226zm-9.542 125.372l32.434 4.558 2.226-15.844-32.433-4.558-2.227 15.844zm25.625-4.477l-5.698 40.541 15.844 2.227 5.698-40.542-15.844-2.226zm3.338 33.732l-32.434-4.558-2.226 15.844 32.433 4.559 2.227-15.845zm-41.469 2.251l-5.698 40.541 15.844 2.227 5.698-40.541-15.844-2.227zm3.338 33.733l-54.291-7.63-2.227 15.844 54.291 7.63 2.227-15.844zm-98.707-80.4l54.29 7.63 2.227-15.844-54.29-7.63-2.227 15.844zm63.304.962l12.558-79.123-15.802-2.508-12.558 79.123 15.802 2.508zm5.771-88.299l-1.058-.148-2.227 15.844 1.058.149 2.227-15.845zm-8.145 2.453a991.73 991.73 0 00-9.061 10.323l12.092 10.477a988.987 988.987 0 018.917-10.158l-11.948-10.642zm-8.912 10.156a721.727 721.727 0 01-8.792 9.43l11.599 11.022a738.52 738.52 0 008.987-9.64L113.095 53.4zm-8.869 9.513l-47.912 51.864 11.752 10.857 47.913-51.864-11.753-10.857z"
+        fill="url(#prefix__paint0_linear)"
+      />
+      <defs>
+        <linearGradient
+          id="prefix__paint0_linear"
+          x1={148.012}
+          y1={-142.526}
+          x2={83.575}
+          y2={315.968}
+          gradientUnits="userSpaceOnUse"
+        >
+          <stop stopColor="#56CCF2" />
+          <stop offset={1} stopColor="#6FCF97" />
+        </linearGradient>
+      </defs>
+    </svg>
+  );
 }
