+++ assets/icons/pencil.tsx
@@ -1,28 +1,26 @@
 export default function Pencil(): JSX.Element {
   return (
-    <>
-      <svg
-        width="20"
-        height="24"
-        viewBox="0 0 24 25"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <path
-          d="M12 20.5H21"
-          stroke="#5F5F5F"
-          strokeWidth="2"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-        />
-        <path
-          d="M16.5 3.99998C16.8978 3.60216 17.4374 3.37866 18 3.37866C18.2786 3.37866 18.5544 3.43353 18.8118 3.54014C19.0692 3.64674 19.303 3.803 19.5 3.99998C19.697 4.19697 19.8532 4.43082 19.9598 4.68819C20.0665 4.94556 20.1213 5.22141 20.1213 5.49998C20.1213 5.77856 20.0665 6.05441 19.9598 6.31178C19.8532 6.56915 19.697 6.803 19.5 6.99998L7 19.5L3 20.5L4 16.5L16.5 3.99998Z"
-          stroke="#5F5F5F"
-          strokeWidth="2"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-        />
-      </svg>
-    </>
+    <svg
+      width="20"
+      height="24"
+      viewBox="0 0 24 25"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M12 20.5H21"
+        stroke="#5F5F5F"
+        strokeWidth="2"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+      <path
+        d="M16.5 3.99998C16.8978 3.60216 17.4374 3.37866 18 3.37866C18.2786 3.37866 18.5544 3.43353 18.8118 3.54014C19.0692 3.64674 19.303 3.803 19.5 3.99998C19.697 4.19697 19.8532 4.43082 19.9598 4.68819C20.0665 4.94556 20.1213 5.22141 20.1213 5.49998C20.1213 5.77856 20.0665 6.05441 19.9598 6.31178C19.8532 6.56915 19.697 6.803 19.5 6.99998L7 19.5L3 20.5L4 16.5L16.5 3.99998Z"
+        stroke="#5F5F5F"
+        strokeWidth="2"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </svg>
   );
 }
