--- assets/icons/spike.tsx
@@ -1,21 +1,0 @@
-export default function Spike(): JSX.Element {
-  return (
-    <>
-      <svg
-        width="20"
-        height="24"
-        viewBox="0 0 22 21"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <path
-          d="M21 10.5H17L14 19.5L8 1.5L5 10.5H1"
-          stroke="black"
-          strokeWidth="2"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-        />
-      </svg>
-    </>
-  );
-}
