+++ components/publicPage/linkCard.tsx
@@ -2,7 +2,7 @@ import Slide from "react-reveal/Slide";
 import React from "react";
 
 import { kzillaxyzdomain } from "../../utils/constants";
-import { truncateTitleText } from "../../utils/functions";
+import { truncateText } from "../../utils/functions";
 
 interface LinkCardProps {
     title: string;
@@ -17,7 +17,7 @@ const LinkCard = ({ title, shortCode, image }: LinkCardProps): JSX.Element => {
                 <a href={`${kzillaxyzdomain}${shortCode}`} className="px-10" target="_blank" rel="noopener noreferrer">
                     <img className="mx-auto -mb-4 rounded" width="55" height="55" src={image} />
                     <div className="bg-white cursor-pointer shadow-custom rounded-2xl text-center">
-                        <h1 className="text-2xl font-extrabold py-6">{truncateTitleText(title)}</h1>
+                        <h1 className="text-2xl font-extrabold py-6">{truncateText(title, 20, 20)}</h1>
                     </div>
                 </a>
             </Slide>
