+++ components/shared/index.tsx
@@ -1,2 +1,3 @@
 export { default as Footer } from "./footer";
 export { default as Navbar } from "./navbar";
+export { default as FloatingCard } from "./floatingcard";
