+++ styles/globals.css
@@ -2,9 +2,6 @@
 @tailwind components;
 @tailwind utilities;
 
-/* TODO:To be removed - Work Sans Noto Sans from css and tailwind config*/
-@import url("https://fonts.googleapis.com/css2?family=Work+Sans&display=swap");
-@import url("https://fonts.googleapis.com/css2?family=Noto+Sans&display=swap");
 @import url('https://fonts.googleapis.com/css2?family=Mulish:wght@400;600;700;900&display=swap');
 
 * {
